requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
redis==4.5.5
passlib[bcrypt]==1.7.5
PyJWT==2.8.0
